+++ spx-gui/src/components/ui/UITag.vue
@@ -25,15 +25,18 @@ defineProps<{
   border-radius: var(--ui-border-radius-2);
   cursor: pointer;
   white-space: nowrap;
+  transition: .3s;
 
   &.type-primary {
+    color: var(--ui-color-grey-100);
     background: var(--ui-color-primary-main);
     border-color: var(--ui-color-primary-main);
   }
 
   &.type-boring {
-    background: var(--ui-color-grey-600);
-    border-color: var(--ui-color-grey-600);
+    color: var(--ui-color-grey-900);
+    background: var(--ui-color-grey-300);
+    border-color: var(--ui-color-grey-300);
   }
 }
 </style>
