+++ spx-gui/src/components/ui/index.ts
@@ -42,6 +42,4 @@ export { default as UIEditorBackdropItem } from './block-items/UIEditorBackdropI
 export { default as UIBackdropItem } from './block-items/UIBackdropItem.vue'
 export { default as UIEditorWidgetItem } from './block-items/UIEditorWidgetItem.vue'
 export { default as UIPagination } from './UIPagination.vue'
-export { default as UIButtonRadioGroup } from './button-radio/UIButtonRadioGroup.vue'
-export { default as UIButtonRadio } from './button-radio/UIButtonRadio.vue'
 export * from './select'
