+++ spx-gui/src/components/ui/radio/index.ts
@@ -1,2 +1,5 @@
 export { default as UIRadio } from './UIRadio.vue'
 export { default as UIRadioGroup } from './UIRadioGroup.vue'
+// TODO: consider merging `UITagRadio` with `UIRadio` (with prop `style="tag"`)
+export { default as UITagRadio } from './UITagRadio.vue'
+export { default as UITagRadioGroup } from './UITagRadioGroup.vue'
